TASK-6: Decomposed JsonShops.cs
