+++ TASK-6/JsonShops.cs
@@ -5,82 +5,6 @@ namespace TASK_6;
 
 [Serializable]
 public class JsonShops
-{
-    public class Order
-    {
-        private static Logger _logger = LogManager.GetCurrentClassLogger();
-        private Guid _id;
-        private Phone _phone;
-        private Shop _shop;
-
-        public Order(Phone phone, Shop shop)
-        {
-            Id = Guid.NewGuid();
-            Phone = phone;
-            Shop = shop;
-            _logger.Info($"Заказ {Phone.Model} на сумму {Phone.Price} успешно оформлен!");
-        }
-        
-        public Guid Id { get => _id; set => _id = value; }
-        public Phone Phone { get => _phone; set => _phone = value; }
-        public Shop Shop { get => _shop; set => _shop = value; }
-    }
-
-    public class Phone
-    {
-        private static Logger _logger = LogManager.GetCurrentClassLogger();
-        private string _model;
-        private string _operationSystemType;
-        private string _marketLaunchDate;
-        private string _price;
-        private bool _isAvailable;
-        private int _shopId;
-
-        public string Model { get => _model; set => _model = value; }
-        public string OperationSystemType { get => _operationSystemType; set => _operationSystemType = value; }
-        public string MarketLaunchDate { get => _marketLaunchDate; set => _marketLaunchDate = value; }
-        public string Price { get => _price; set => _price = value; }
-        public bool IsAvailable { get => _isAvailable; set => _isAvailable = value; }
-        public int ShopId { get => _shopId; set => _shopId = value; }
-
-        public void Report()
-        {
-            _logger.Info(
-                $"Модель: {Model}, тип ОС: {OperationSystemType}, дата выпуска: {MarketLaunchDate}, " +
-                $"стоимость: {Price}, магазин: {Program.Shops.ShopByNumber(ShopId)}\n");
-        }
-    }
-
-    public class Shop
-    {
-        private static Logger _logger = LogManager.GetCurrentClassLogger();
-        private int _id;
-        private string _name;
-        private string _description;
-        private List<Phone> _phones;
-
-        public int Id { get => _id; set => _id = value; }
-        public string Name { get => _name; set => _name = value; }
-        public string Description { get => _description; set => _description = value; }
-        public List<Phone> Phones { get => _phones; set => _phones = value; }
-
-        public int CountPhonesWithOsType(string osType)
-        {
-            var foundPhones = Phones.FindAll(p => p.OperationSystemType == osType && p.IsAvailable);
-            var numberOfPhones = foundPhones.Count();
-
-            return numberOfPhones;
-        }
-
-        public void Report()
-        {
-            _logger.Info(
-                "\nМагазин №{0} {1} - {2}.\nКоличество устройств IOS в наличии: {3}\nКоличество устройств Android в наличии: {4}\n",
-                Id, Name, Description, CountPhonesWithOsType("IOS"), CountPhonesWithOsType("Android"));
-        }
-    }
-
-    public class AllShops
     {
         private static Logger _logger = LogManager.GetCurrentClassLogger();
         private List<Shop> _shops;
@@ -175,4 +99,3 @@ public class JsonShops
             return phoneChoice;
         }
     }
-}
