+++ TASK-6/Program.cs
@@ -2,13 +2,13 @@
 
 namespace TASK_6;
 
-public class Program
+public static class Program
 {
-    private static JsonShops.AllShops _shops;
-    private static List<JsonShops.Phone> _foundPhones;
+    private static JsonShops _shops;
+    private static List<Phone> _foundPhones;
 
-    public static JsonShops.AllShops Shops { get => _shops; set => _shops = value; }
-    public static List<JsonShops.Phone> FoundPhones { get => _foundPhones; set => _foundPhones = value; }
+    public static JsonShops JsonShops { get => _shops; set => _shops = value; }
+    public static List<Phone> FoundPhones { get => _foundPhones; set => _foundPhones = value; }
 
     public static void Main(string[] args)
     {
@@ -17,20 +17,20 @@ public class Program
         {
             json = streamReader.ReadToEnd();
         }
-
-        Shops = JsonConvert.DeserializeObject<JsonShops.AllShops>(json);
+        
+        JsonShops = JsonConvert.DeserializeObject<JsonShops>(json);
         Console.WriteLine();
 
-        foreach (var shop in Shops.Shops)
+        foreach (var shop in JsonShops.Shops)
         {
             shop.Report();
         }
 
-        var chosenPhone = Shops.FindPhone();
-        var chosenShop = Shops.ChooseShop(chosenPhone);
+        var chosenPhone = JsonShops.FindPhone();
+        var chosenShop = JsonShops.ChooseShop(chosenPhone);
         var selectedPhone = FoundPhones.Find(p => p.Model == chosenPhone && p.ShopId == chosenShop.Id);
 
-        var order = new JsonShops.Order(selectedPhone, chosenShop);
+        var order = new Order(selectedPhone, chosenShop);
         var json2 = JsonConvert.SerializeObject(order);
         using (StreamWriter file = File.CreateText(@"D:\order.txt"))
         {
